refactor(structure): invites + meetings + program + sections layout
- invites:
  * useWardInvites → hooks/
  * inviteActions, inviteMailto → utils/
  * inviteFormField.tsx → InviteField.tsx (PascalCase to match the
    primary exported component)
- meetings (root):
  * 10 utils → utils/
  * 6 util tests → utils/__tests__/
- meetings/program:
  * useApprovalActions → hooks/
  * railSections + test → utils/ + utils/__tests__/
  * fix-up: useApprovalActions moved one level deeper, adjust its
    "../approvals" → "../../utils/approvals"
- meetings/sections:
  * speakerListItems + test → utils/ + utils/__tests__/

External importers updated where applicable. Components stay flat.
Typecheck + 346 tests + build all pass.
