+++ src/features/meetings/program/ProgramHead.tsx
@@ -1,6 +1,6 @@
 import { Link } from "react-router";
 import type { MeetingType } from "@/lib/types";
-import { TYPE_LABELS } from "../meetingLabels";
+import { TYPE_LABELS } from "../utils/meetingLabels";
 import { cn } from "@/lib/cn";
 
 interface Props {
