+++ src/features/meetings/sections/utils/speakerListItems.ts
