+++ src/features/meetings/utils/readiness.ts
