+++ src/features/speakers/SpeakerForm.tsx
@@ -1,7 +1,7 @@
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
-import { ensureMeetingDoc } from "@/features/meetings/ensureMeetingDoc";
+import { ensureMeetingDoc } from "@/features/meetings/utils/ensureMeetingDoc";
 import type { NonMeetingSunday } from "@/lib/types";
 import { createSpeaker } from "./speakerActions";
