+++ src/features/speakers/speakerActions.ts
@@ -1,7 +1,7 @@
 import { collection, doc, serverTimestamp, writeBatch } from "firebase/firestore";
-import { writeMeetingPatch } from "@/features/meetings/approvals";
-import { ensureMeetingDoc } from "@/features/meetings/ensureMeetingDoc";
-import { appendHistoryEvent, currentActor } from "@/features/meetings/history";
+import { writeMeetingPatch } from "@/features/meetings/utils/approvals";
+import { ensureMeetingDoc } from "@/features/meetings/utils/ensureMeetingDoc";
+import { appendHistoryEvent, currentActor } from "@/features/meetings/utils/history";
 import { db } from "@/lib/firebase";
 import { reportSaved, reportSaveError, reportSaving } from "@/stores/saveStatusStore";
 import type { HistoryChange, NonMeetingSunday, SpeakerRole, SpeakerStatus } from "@/lib/types";
